Исправления, Модуль Product

+++ app/Http/Controllers/Web/CategoryController.php
@@ -5,47 +5,81 @@ namespace App\Http\Controllers\Web;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Web\CategoryRequest;
 use App\Models\Category;
+use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Http\Request;
+
+// Добавляем пространство имен для политики
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class CategoryController extends Controller
 {
+    // Добавляем трейд для политики
+    use AuthorizesRequests;
+    /**
+     * Display a listing of the resource.
+     */
     public function index()
     {
         $categories = Category::all();
-
         return view('categories.index', compact('categories'));
     }
 
+    /**
+     * Show the form for creating a new resource.
+     */
     public function create()
     {
+        $this->authorize('create', Category::class);
         return view('categories.create');
     }
 
+    /**
+     * Store a newly created resource in storage.
+     */
     public function store(CategoryRequest $request)
     {
+        $this->authorize('create', Category::class);
+
         // Создание категории
         $category = Category::create($request->validated());
         return redirect()->route('categories.index');
     }
 
+    /**
+     * Display the specified resource.
+     */
     public function show(Category $category)
     {
         return view('categories.show', compact('category'));
     }
 
+    /**
+     * Show the form for editing the specified resource.
+     */
     public function edit(Category $category)
     {
+        $this->authorize('update', $category);
         return view('categories.edit', compact('category'));
     }
 
+    /**
+     * Update the specified resource in storage.
+     */
     public function update(CategoryRequest $request, Category $category)
     {
+        $this->authorize('update', $category);
+
         $category->update($request->validated());
         return redirect()->route('categories.index');
     }
 
+    /**
+     * Remove the specified resource from storage.
+     */
     public function destroy(Category $category)
     {
+        $this->authorize('delete', $category);
+
         $category->delete();
         return redirect()->route('categories.index');
     }
